// backend/src/modules/tramites/domain/enums/estado-tramite.enum.ts
export enum EstadoTramite {
  BORRADOR = 'BORRADOR',
  INGRESADO = 'INGRESADO',
  EN_REVISION = 'EN_REVISION',
  OBSERVADO = 'OBSERVADO',
  DERIVADO = 'DERIVADO',
  ESPERANDO_EXTERNO = 'ESPERANDO_EXTERNO',
  ESPERANDO_INTERNO = 'ESPERANDO_INTERNO',
  APROBADO = 'APROBADO',
  RECHAZADO = 'RECHAZADO',
  CANCELADO = 'CANCELADO',
  CERRADO = 'CERRADO',
}