// backend/src/modules/tramites/domain/enums/accion-workflow.enum.ts
export enum AccionWorkflow {
  CREAR = 'CREAR',
  INGRESAR = 'INGRESAR',
  TOMAR = 'TOMAR',
  ASIGNAR = 'ASIGNAR',
  DERIVAR = 'DERIVAR',
  OBSERVAR = 'OBSERVAR',
  RESPONDER_OBSERVACION = 'RESPONDER_OBSERVACION',
  SOLICITAR_INTERVENCION_EXTERNA = 'SOLICITAR_INTERVENCION_EXTERNA',
  RESPONDER_INTERVENCION_EXTERNA = 'RESPONDER_INTERVENCION_EXTERNA',
  APROBAR = 'APROBAR',
  RECHAZAR = 'RECHAZAR',
  CANCELAR = 'CANCELAR',
  CERRAR = 'CERRAR',
}